Extract scrypt module
